p2p/dht:update space limit cache

+++ system/p2p/dht/protocol/broadcast/block.go
@@ -166,7 +166,7 @@ func (protocol *broadCastProtocol) recvLtBlock(ltBlock *types.LightBlock, pid, p
 	}
 
 	//需要将不完整的block预存
-	protocol.ltBlockCache.Add(blockHash, block, int64(block.Size()))
+	protocol.ltBlockCache.Add(blockHash, block, block.Size())
 	//pub to specified peer
 	_, err := protocol.sendPeer(pid, query, false)
 	if err != nil {
